+++ app/components/viral/breadcrumb_component.html.erb
