+++ app/components/viral/flash_component.html.erb
