+++ app/helpers/view_helper.rb
@@ -3,6 +3,7 @@
 # ViewHelper for user interface components
 module ViewHelper
   VIRAL_HELPERS = {
+    flash: 'Viral::FlashComponent',
     icon: 'Viral::IconComponent'
   }.freeze
 
@@ -10,18 +11,6 @@ module ViewHelper
     define_method "viral_#{name}" do |*args, **kwargs, &block|
       render component.constantize.new(*args, **kwargs), &block
     end
-  end
-
-  # Get the svg file for a heroicon by name, and add the applicable classes
-  def heroicons_source(icon_name, classes)
-    path = Rails.root.join('app', 'assets', 'icons', 'heroicons', "#{icon_name}.svg")
-    file = File.read(path)
-    doc = Nokogiri::HTML::DocumentFragment.parse(file)
-    svg = doc.at_css 'svg'
-    svg[:class] = classes
-    svg[:focusable] = false
-    svg[:'aria-hidden'] = true
-    doc.to_html.html_safe # rubocop:disable Rails/OutputSafety
   end
 
   def viral_icon_source(name)
