--- app/views/shared/_flash.html.erb
@@ -1,5 +1,0 @@
-<div class="fixed top-5 right-6 flex flex-col ">
-  <% flash.each do |key, value| %>
-    <%= render(FlashComponent.new(type: key, data: value)) %>
-  <% end %>
-</div>
